worker/requirements.txt
python-dotenv==1.0.1
psycopg[binary]==3.2.4
httpx==0.27.2
PyJWT==2.9.0

